Scheduled Revisions: "Reload" link was broken by output escaping
This link is displayed only for scheduled revisions previewed shortly after creation.  Before the next release, we'll review whether the new scheduling mechanism eliminates the need for this workaround.

In the past, there were cases (probably due to caching performance on the server) where the page template and post thumbnail were not displayed when the preview is accessed immediately after creation.